+++ h_main.cu
@@ -8,10 +8,10 @@
 #include "molecule.h"
 #include "h_main.h"
 
-void discombob(float * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms);
+void discombob(double * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms);
 
 
-int discombob_on_cpu(float * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms){
+int discombob_on_cpu(double * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms){
 
     cudaEvent_t start_cpu, stop_cpu;
     float cpuMsecTime = -1;
@@ -28,7 +28,7 @@ int discombob_on_cpu(float * energyGrid, atom *atoms, int dimX, int dimY, int di
     return cpuMsecTime;
 }
 
-void discombob(float * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms) {
+void discombob(double * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms) {
     int i,j,k,n;
     for (k=0; k<dimZ; k++) {
         float z = gridSpacing * (float)k;
@@ -36,12 +36,12 @@ void discombob(float * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, fl
             float y = gridSpacing * (float)j;
             for (i = 0; i < dimX; i++){
                 float x = gridSpacing * (float)i;
-                float energy = 0.0f;
+                double energy = 0.0;
                 for (n = 0; n<numAtoms; n++){
-                    float dx = x - atoms[n].x;
-                    float dy = y - atoms[n].y;
-                    float dz = z - atoms[n].z;
-                    float charge = atoms[n].charge;
+                    double dx = (double)x - atoms[n].x;
+                    double dy = (double)y - atoms[n].y;
+                    double dz = (double)z - atoms[n].z;
+                    double charge = atoms[n].charge;
                     energy += charge/sqrt(dx*dx + dy*dy + dz*dz);
                 }
                 energyGrid[dimX*dimY*k + dimX*j + i] = energy;
