+++ h_main.h
@@ -1,1 +1,1 @@
-int discombob_on_cpu(float * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms);
+int discombob_on_cpu(double * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing, int numAtoms);
